+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support for the `incremental` profile attribute.
 - Set all performance-related default properties from the base profile (dev/release) on
   templates (https://github.com/Kobzol/cargo-wizard/issues/4).
+- Add support for the `split-debuginfo` profile attribute.
 
 # 0.2.1 (10. 3. 2024)